update the course project
