+++ courseProject/constants.js
@@ -1,0 +1,2 @@
+export const apiKey = 'SG5l9cClWjV6Ml3gsTrhBMGZ2JwH8Om0';
+export const apiUrl = "https://calendarific.com/api/v2";
