+++ courseProject/index.html
@@ -18,17 +18,16 @@
 
         <div class="content">
             <div class="tab-content">
-                <h3 id="text">Будь ласка, виберіть початкову та кінцеву дату.</h3>
+                <h3 id="text">Будь ласка, виберіть початкову дату.</h3>
                 <label for="startDate">Початкова дата:</label>
                 <input type="date" id="startDate" name="startDate">
 
                 <label for="datePreset">Пресет часу:</label>
-                <select id="datePreset" name="datePreset">
-                    <option value="week">Вибрати вручну</option>
+                <select id="datePreset" name="datePreset" disabled>
+                    <option value="manual">Вибрати вручну</option>
                     <option value="week">Тиждень</option>
                     <option value="month">Місяць</option>
                 </select>
-
                 <label for="endDate">Кінцева дата:</label>
                 <input type="date" id="endDate" name="endDate" disabled min="" max="">
 
@@ -47,15 +46,15 @@
                     <option value="seconds">Порахувати кількість секунд</option>
                 </select>
 
-                <button id="getDateRangeButton">Отримати проміжок дат</button>
-
+                <div>
+                    <button id="getDateRangeButton">Отримати проміжок дат</button>
+                </div>
                 <div id="dateRangeResult"></div>
 
                 <div id="lastResultsTable">
                     <h3>Останні результати</h3>
                     <button id="showLastResultsButton">Останні результати</button>
                     <table>
-                        <!-- <caption>Результати роозрахунків</caption> -->
                         <thead>
                             <tr>
                                 <th>Початкова дата</th>
@@ -77,22 +76,27 @@
                 <select id="year" disabled></select>
 
                 <button id="getHolidaysButton">Отримати свята</button>
-                <!-- <table>
-                    <thead>
-                        <tr>
-                            <th>Дата</th>
-                            <th>Назва свята</th>
-                        </tr>
-                    </thead> -->
-                <tbody id="lastResultsTableBodyHolidays"></tbody>
-                <!-- </table> -->
+                <div id="holidaysResult">
+                    <table>
+                        <thead>
+                            <tr>
+                                <th>
+                                    <button id="sortDateButton"
+                                        onclick="window.holidaysTab.sortTable('lastResultsTableBodyHolidays', 0)">
+                                        Дата<img id="rotateImage" class="img" src="./Vector.svg">
+                                    </button>
 
-                <div id="holidaysResult"></div>
+                                </th>
+                                <th>Назва свята</th>
+                            </tr>
+                        </thead>
+                        <tbody id="lastResultsTableBodyHolidays"></tbody>
+                    </table>
+                </div>
             </div>
         </div>
     </div>
-
-    <script src="./script.js"></script>
+    <script type="module" src="main.js"></script>
 </body>
 
 </html>
