+++ courseProject/styles.css
@@ -4,8 +4,6 @@
 
 body {
     font-family: Arial, sans-serif;
-    margin: 0;
-    padding: 0;
     display: flex;
     justify-content: center;
     align-items: center;
@@ -24,7 +22,7 @@ input[type="radio"] {
 }
 
 label {
-    padding: 10px 20px;
+    padding: 10px 50px;
     display: inline-block;
     cursor: pointer;
     background-color: #ddd;
@@ -33,11 +31,12 @@ label {
 }
 
 label:hover {
-    background-color: #eee;
+    background-color: #e7e5e5;
 }
 
-.content {
-    display: flex;
+#tab1:checked+label.active,
+#tab2:checked+label.active {
+    background-color: #89f3d5;
 }
 
 .tab-content {
@@ -54,7 +53,7 @@ button {
 }
 
 table {
-    width: 80%;
+    width: 100%;
 }
 
 table,
@@ -71,20 +70,43 @@ td {
     font-size: 15px;
 }
 
-
-
 #showLastResultsButton,
 #getDateRangeButton,
 #getHolidaysButton {
     font-size: 16px;
     padding: 10px 15px;
-    background-color: #e89cf6;
+    background-color: #169e7e;
     color: #fff;
     border: none;
     cursor: pointer;
 }
 
+#sortDateButton {
+    font-family: fantasy, sans-serif;
+    display: inline-flex;
+    justify-content: center;
+    align-items: center;
+    background-color: #f3f3f3;
+    font-size: 16px;
+    padding: 4px 25px;
+    border: none;
+    cursor: pointer;
+    font-weight: bold;
+}
+
+.img {
+    margin-left: 5px;
+    transition: transform 0.5s ease-in-out;
+    --rotation: 0deg;
+    transform: rotate(var(--rotation));
+}
+
+
 @media screen and (max-width: 380px) {
+    label {
+        padding: 10px 40px;
+    }
+
     th,
     td {
         font-size: 10px;
